+++ BottangoArduinoDriver/src/Services/I2SHelper.h
@@ -13,7 +13,7 @@
 #endif
 
 #include "../../BottangoArduinoConfig.h"
-#include "../Data Source/SDCardUtil.h"
+#include "../DataSource/SDCardUtil.h"
 
 namespace I2SHelper
 {
